% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_grid.R
\name{circle_grid}
\alias{circle_grid}
\title{Circle grid}
\usage{
circle_grid(center_x = 50, center_y = 50, radius = 100, spacing_x = 7,
  spacing_y = 7, show_plot = TRUE)
}
\arguments{
\item{center_x}{Grid's center X coordinate.}

\item{center_y}{Grid's center Y coordinate.}

\item{radius}{Radius of receptor grid.}

\item{spacing_x}{Distance between receptors in the x direction.}

\item{spacing_y}{Distance between receptors in the y direction.}

\item{show_plot}{Plot receptor grid.}
}
\description{
Generate a circular receptor grid
}
\examples{
circle_grid(center_x   = 0,
            center_y   = 0,
            radius     = 100,
            spacing_x  = 7,
            spacing_y  = 7)
}
\keyword{circle}
\keyword{grid}
\keyword{receptors}

