% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_grid.R
\name{polar_grid}
\alias{polar_grid}
\title{Polar grid}
\usage{
polar_grid(center_x = 100, center_y = 100, radii = c(10, 15, 20),
  n_radials = 36, show_plot = TRUE)
}
\arguments{
\item{center_x}{Grid's center X coordinate.}

\item{center_y}{Grid's center Y coordinate.}

\item{radii}{Distances of receptor circles from center.}

\item{n_radials}{Number of radials / spokes of receptors.}

\item{show_plot}{Plot receptor grid.}
}
\description{
Generate a polar receptor grid
}
\examples{
polar_grid(center_x = 0,
           center_y = 0,
           length = 1,
           width = 2)
}
\keyword{grid}
\keyword{polar}
\keyword{receptors}

