% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rou.R
\name{read_rou}
\alias{read_rou}
\title{Read an AERMOD .rou receptor file into a data frame}
\usage{
read_rou(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or a text string.}
}
\description{
Read an AERMOD .rou receptor file into a data frame of receptor coordinates.
}
\examples{
\dontrun{
read_rou(file = "circle_receptors.rou")
}
}
\keyword{aermod}
\keyword{read}
\keyword{receptors}
\keyword{rou}

