% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect_grid.R
\name{rect_grid}
\alias{rect_grid}
\title{Rectangle grid}
\usage{
rect_grid(center_x = 50, center_y = 50, length = 100, width = 200,
  spacing_x = 7, spacing_y = 7, show_plot = TRUE)
}
\arguments{
\item{center_x}{Grid's center X coordinate.}

\item{center_y}{Grid's center Y coordinate.}

\item{length}{Length of receptor grid in the x direction.}

\item{width}{Width of receptor grid in the y direction.}

\item{spacing_x}{Distance between receptors in the x direction.}

\item{spacing_y}{Distance between receptors in the y direction.}

\item{show_plot}{Plot receptor grid.}
}
\description{
Generate a rectangular receptor grid
}
\examples{
rect_grid(center_x = 0,
          center_y = 0,
          length   = 100,
          width    = 200,
          spacing_x  = 7,
          spacing_y  = 7)
}
\keyword{grid}
\keyword{receptors}
\keyword{rectangle}

