% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_grid.R
\name{ring_grid}
\alias{ring_grid}
\title{Ring grid}
\usage{
ring_grid(center_x = 100, center_y = 100, radii = c(10, 15, 20, 30, 100),
  spacing = 7, show_plot = TRUE)
}
\arguments{
\item{center_x}{Grid's center X coordinate.}

\item{center_y}{Grid's center Y coordinate.}

\item{radii}{Distances of receptor rings from center.}

\item{spacing}{Distance between receptors. A single value will apply to all rings. Enter a sequence of values to apply a unique distance for each ring.}

\item{show_plot}{Plot receptor grid.}
}
\description{
Generate rings of receptors at increasing distances from a center point.
}
\examples{
ring_grid(center_x = 0,
          center_y = 0,
          radii    = c(10,15,20,30,100),
          spacing  = 7)

}
\keyword{grid}
\keyword{receptors}
\keyword{ring}

