% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rings_grid.R
\name{rings_grid}
\alias{rings_grid}
\title{Rings grid}
\usage{
rings_grid(center_x = 0, center_y = 0, radii = c(10, 15, 20, 30, 100),
  spacing = 7, show_plot = TRUE)
}
\arguments{
\item{center_x}{Grid's center X coordinate.}

\item{center_y}{Grid's center Y coordinate.}

\item{radii}{Distances of receptor rings from center.}

\item{spacing}{Distance between receptors. A single value will apply to all rings. Enter a vector of values to apply a unique spacing distance for each ring.}

\item{show_plot}{Plot receptor grid.}
}
\description{
Generate rings of receptors at increasing distances from a center point.
}
\examples{
rings_grid(center_x = 0,
           center_y = 0,
           radii    = c(10,15,20,30,100),
           spacing  = 7)

}
\keyword{grid}
\keyword{receptors}
\keyword{rings}

