% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rou.R
\name{write_rou}
\alias{write_rou}
\title{Write .rou receptor file}
\usage{
write_rou(data, path = NULL, grid_type = "disccart",
  elev_units = "meters", ...)
}
\arguments{
\item{data}{Dataframe of receptor coordinates and elevations. Requires columns: `x`, `y`, and `elevation`.}

\item{path}{Path to write to. If NULL, returns result as text.}

\item{grid_type}{Type of receptor grid for AERMOD. Default is "disccart" (discrete cartesian).}

\item{elev_units}{Units for elevation. Default is "meters".}
}
\description{
Write a .rou receptor file for AERMOD from a dataframe of receptor coordinates.
}
\examples{
receptors <- circle_grid(center_x = 50, center_y= 50, radius = 100)

receptors$elevation <- 250

write_rou(data = receptors,
          path = "circle_receptors.rou")
}
\keyword{aermod}
\keyword{receptors}
\keyword{rou}
\keyword{save}
\keyword{write}

